requirements.txt
beautifulsoup4==4.9.2
Flask==1.1.1
gunicorn==20.0.4
mysql-connector-python==8.0.16
pylint==2.5.2
pylint-exit==1.1.0
requests==2.24.0
urllib3==1.24.3